Added helper methods for validating and creating payloads for pool create. Moved things around a bit.
